Add instance Arbitrary Block
And adjust the instance Arbitrary BlockHeader

Move the genChainAnchor helper into the ChainGenerators module from the
ChainFragment module.
